More whitespace ignoring, Error messages
Ignores whitespace after an open parenthesis. Due to a combination of this, and the fact that the system uses the multiplication operator '*' as a wildcard, this is only supported if the equation has quotes around it on the command line. Added error checking to help check if this is the case and print a descriptive error.